remove bitcoinj dependency
